Pull out object for mapping entities to XPath queries
